Add DiscordBot class and integrate into main application
A new class called `DiscordBot.cs` was Integrated into the main application to handle discord interactions. This includes event implementations for receiving, processing, and replying users' messages. Also included is a readiness check to notify dev/admin when the bot is online.

Added Discord package reference to `StoryTeller.csproj` to support the new Discord bot class.

In `.gitignore`, included `token.txt` due to security implications of accidentally committing the token.

Additionally, a minor alteration was done in `Program.cs` to launch the discord bot once option 3 is chosen and changed the path of `modelPath`.

This update allows the application to handle discord messaging, broadening the usage scope of the software.