Update bot modes and add new functionalities
This commit refactors and separates the modes of operation for the bot. Removed the original DiscordBot.cs and added: DiscordBot.cs, StoryTellerBot.cs, ChatBot.cs under a new BotModes directory. Also, introduced an interface, IMode.cs for these modes to operate under. In addition, added Session handling with Sessions.cs for better management of user states and inputs. Write token.txt to gitignore to prevent accidental pushes and a Version folder to keep track of application iterations.

Reasons for these changes are to improve the organization of the code and enhance its maintainability. It also aims to enhance user experience by introducing session handling which allows the bot to remember past interactions. Also, dividing the bot's operation mode aids customizability for future improvements or additions.