Refactor code to use var keyword
In this commit, explicit variable declarations were changed to use 'var' keyword to improve readability and maintainability of the code. Following best practices and recommendations from Microsoft, 'var' keyword was used in place of explicit types where the assignment statement clearly implies the type of the variable.

Reduced redundancy and potential errors, by inferring the type of the variable from the initialization expression. This change spans across multiple files, including but not limited to DataLoader.cs, Sessions.cs, BotModes files, etc.

Also, in the ChatBot.cs, instead of throwing an exception, bot's chat method is invoked to follow logical consistency. Regarding Modes in Sessions.cs, the '?.' operator was utilized for null-checking to make the code safer.