Refactor CreateSession function and remove modelPath argument
Removed 'modelPath' as an argument in the 'CreateSession' function in Sessions.cs. The model path is now hard-coded within the 'CreateSession' method itself, eliminating the need for it to be passed as an argument each time the function is called. Therefore, the corresponding calls in Program.cs were updated to reflect this change. This was done to simplify the session creation process.