Remove StoryTeller version 0.1 for clean up
StoryTeller version 0.1 was deleted as part of a general clean-up process. This version has become obsolete and is no longer in use, furthermore, it was causing confusion and potential conflicts with the most current and stable version. Additionally, a new Jupyter notebook 'RunTempTest.ipynb' was created in the 'NoteBooks' directory. This notebook runs various scenarios on different temperature settings and generates a report of the results. This feature will help in the improvement of temperature-dependent functions and evaluating their performance across various settings.