Add Category and RootObject classes in Category.cs
Created two new classes, Category and RootObject, in the Category.cs file in the Classes directory. Each class has a property corresponding to a list of strings. Also, script for running temperature tests in RunTempTestV2 notebook was added. This was necessary to prepare the application for future updates in handling of categories. The RunTempTest notebook contains further information for testing the features added.